+++ src/RealPolynomial.h
@@ -4,12 +4,12 @@ class RealPolynomial
 {
 private:
     int n{};
-    std::vector<float> element{};
+    std::vector<double> element{};
 public:
-    RealPolynomial(std::vector<float> element);
+    RealPolynomial(std::vector<double> element);
     ~RealPolynomial();
     void printPol();
-    float evaluate(float x);
+    double evaluate(double x);
     RealPolynomial* addPolynomial(RealPolynomial other);
 };
